web/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.8.0
socket.io-client==5.8.0
pandas==2.1.3
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
pyarrow==14.0.1
